TAUR-964 Addressed Austin's code review feedback.
TAUR-964 Fixed spelling of Heinz in metrics config (was misspelled as Heiz)
