+++ taurus.metric_collectors/conf/application.conf
@@ -2,7 +2,7 @@
 
 
 
-[company_maintenance_agent]
+[metric_maintenance_agent]
 
 # Operation mode; see taurus.metric_collectors.ApplicationConfig for supported
 # mode values
