+++ taurus.metric_collectors/conf/supervisord.conf
@@ -28,4 +28,4 @@ files = ../taurus/metric_collectors/xignite/conf/xignite_security_news_agent.con
         ../taurus/metric_collectors/twitterdirect/conf/twitter_direct_agent.conf
         ../taurus/metric_collectors/twitterdirect/conf/tweet_deletion_agent.conf
         ../taurus/metric_collectors/twitterdirect/conf/twitter_screen_name_checker.conf
-        ../taurus/metric_collectors/common_services/conf/company_maintenance_service.conf
+        ../taurus/metric_collectors/common_services/conf/metric_maintenance_agent.conf
