+++ taurus.metric_collectors/taurus/metric_collectors/common_services/conf/metric_maintenance_agent.conf
@@ -1,6 +1,6 @@
-[program:company_maintenance_agent]
+[program:metric_maintenance_agent]
 directory=%(here)s/../taurus/metric_collectors/common_services
-command=taurus-company-maintenance-agent
+command=taurus-collectors-metric-maintenance-agent
 numprocs=1
 autostart=true
 autorestart=unexpected
@@ -8,5 +8,5 @@ startsecs=0
 exitcodes=0,2
 # Use SIGINT instead of SIGTERM for the builtin KeyboardInterrupt exception
 stopsignal=INT
-stdout_logfile=%(here)s/../logs/company_maintenance_agent.log
+stdout_logfile=%(here)s/../logs/metric_maintenance_agent.log
 redirect_stderr=true
