+++ taurus.metric_collectors/taurus/metric_collectors/metric_csv_archive/metrics.csv
@@ -125,7 +125,7 @@ NSC,NYSE,Norfolk Southern Corp,Norfolk Southern,0.25%,2.1K,$239.9K,@nscorp,,,
 MU,NASDAQ,Micron Technology Inc,Micron Technology,0.24%,7.2K,$237.7K,@MicronTech,,,
 ADBE,NASDAQ,Adobe Systems Inc,Adobe Systems,0.24%,3.4K,$237.5K,@Adobe,,,
 ANTM,NYSE,Anthem Inc,Anthem,0.24%,1.9K,$236.7K,@AnthemInc,,,
-KHC,NASDAQ,Kraft Heinz Co,Kraft Heiz Company,0.24%,4.0K,$233.6K,@KraftFoods,,,
+KHC,NASDAQ,Kraft Heinz Co,Kraft Heinz Company,0.24%,4.0K,$233.6K,@KraftFoods,,,
 ADM,NYSE,Archer-Daniels Midland Co,Archer-Daniels Midland,0.24%,4.4K,$230.4K,@ADMUpdates,,,
 PCP,NYSE,Precision Castparts Corp,Precision Castparts,0.23%,980,$228.2K,,,,
 JCI,NYSE,Johnson Controls Inc,Johnson Controls,0.23%,4.5K,$224.3K,@JohnsonControls,,,
