+++ taurus.metric_collectors/tests/deployment/health_check_test.py
@@ -66,10 +66,10 @@ class TaurusMetricCollectorsHealthCheckTestCase(unittest.TestCase):
 
 
   @_RETRY_SERVICE_RUNNING_CHECK
-  def testCompanyMaintenanceAgentIsRunning(self):
+  def testMetricMaintenanceAgentIsRunning(self):
     # NOTE: This service exits when configured for hot-standby mode
-    self.assertIn(self._getServiceStateName("company_maintenance_agent"),
-                     ["RUNNING", "EXITED"])
+    self.assertIn(self._getServiceStateName("metric_maintenance_agent"),
+                  ["RUNNING", "EXITED"])
 
 
   @_RETRY_SERVICE_RUNNING_CHECK
