+++ taurus.metric_collectors/tests/unit/metric_utils_test.py
@@ -21,7 +21,7 @@
 # ----------------------------------------------------------------------
 
 """
-unit tests for taurus.metric_collectors.metric_utils
+Unit tests for taurus.metric_collectors.metric_utils
 """
 
 # Disable warning "access to protected member" and "method could be a function"
@@ -31,7 +31,6 @@ unit tests for taurus.metric_collectors.metric_utils
 from datetime import datetime, timedelta
 import json
 import logging
-import time
 import unittest
 
 import mock
@@ -39,8 +38,6 @@ from mock import ANY, MagicMock, Mock, patch
 
 
 import sqlalchemy
-
-from nta.utils.test_utils import time_test_utils
 
 from taurus.metric_collectors import metric_utils
 
@@ -318,16 +315,6 @@ class MetricUtilsTestCase(unittest.TestCase):
                      "onlyMetricNames is empty")
 
 
-  def testCreateAllModelsWithDuplicatesInMetricFilter(self):
-    with self.assertRaises(ValueError) as assertContext:
-      metric_utils.createAllModels(host="host",
-                                   apiKey="apikey",
-                                   onlyMetricNames=["a", "a", "b"])
-
-    self.assertEqual(assertContext.exception.args[0],
-                     "onlyMetricNames contains duplicates")
-
-
   def testCreateAllModelsWithUnknownsInMetricFilter(self):
     with self.assertRaises(ValueError) as assertContext:
       metric_utils.createAllModels(host="host",
@@ -372,7 +359,7 @@ class MetricUtilsTestCase(unittest.TestCase):
       "XIGNITE.FOOBAR.CLOSINGPRICE",
       "XIGNITE.FOOBAR.VOLUME",
       "TWITTER.TWEET.HANDLE.FOOBAR.VOLUME",
-      "XIGNTE.NEWS.FOOBAR.VOLUME",
+      "XIGNITE.NEWS.FOOBAR.VOLUME",
     ]
 
     # Execute
@@ -717,342 +704,6 @@ class MetricUtilsTestCase(unittest.TestCase):
                 for m, v, t
                 in samples[metric_utils._METRIC_DATA_BATCH_WRITE_SIZE:]])))
     self.assertEqual(messageBusMock.publish.call_args_list[1], call1)
-
-
-
-class CompanyDeleterTestCase(unittest.TestCase):
-
-
-  @patch("taurus.metric_collectors.metric_utils.collectorsdb.engineFactory",
-         autospec=True)
-  @patch("taurus.metric_collectors.metric_utils.CompanyDeleter"
-         "._flushTaurusEngineMetricDataPath",
-         spec_set=metric_utils.CompanyDeleter._flushTaurusEngineMetricDataPath)
-  @patch("taurus.metric_collectors.metric_utils.getAllMetricSecurities",
-         autospec=True)
-  @patch("taurus.metric_collectors.metric_utils.getAllCustomMetrics",
-         autospec=True)
-  @patch("taurus.metric_collectors.metric_utils.deleteMetric",
-         autospec=True)
-  def testDeleteCompanies(self,
-                          deleteMetricMock,
-                          getAllCustomMetricsMock,
-                          getAllMetricSecuritiesMock,
-                          _flushTaurusEngineMetricDataPathMock,
-                          engineFactoryMock):
-    negatives = set([
-      "TWITTER.TWEET.HANDLE.ZZZ.VOLUME",
-      "XIGNITE.ZZZ.CLOSINGPRICE",
-      "XIGNITE.FOOBARZZZ.CLOSINGPRICE",
-      "XIGNITE.ZZZFOOBAR.CLOSINGPRICE",
-      "XIGNITE.FOOBAR.ZZZ.VOLUME",
-      "XIGNITE.NEWS.FOOBAR.ZZZ.VOLUME",
-      "FOOBAR.VOLUME",
-      ".FOOBAR.CLOSINGPRICE",
-      "XIGNITE.FOOBAR.",
-      "FOOBARCLOSINGPRICE",
-    ])
-
-    positives = set([
-      "XIGNITE.FOOBAR.CLOSINGPRICE",
-      "XIGNITE.FOOBAR.VOLUME",
-      "TWITTER.TWEET.HANDLE.FOOBAR.VOLUME",
-      "XIGNTE.NEWS.FOOBAR.VOLUME",
-      "XIGNITE.DOLITTLE.VOLUME",
-    ])
-
-    # Patch getAllCustomMetrics to return all negatives and positives
-    getAllCustomMetricsMock.return_value = [
-      {"name": metric} for metric in negatives.union(positives)
-    ]
-
-    # Simulate xitgnite_security found for first symbol, but not the second one.
-    engineFactoryMock.return_value.begin.return_value.__enter__.return_value \
-      .execute.side_effect = [Mock(rowcount=1), Mock(rowcount=0)]
-
-    getAllMetricSecuritiesMock.return_value = [
-      ("IBM", "exg"),
-      ("T", "exg")
-    ]
-
-    # Execute the function under test
-    metric_utils.CompanyDeleter.deleteCompanies(
-      tickerSymbols=("FOOBAR", "DOLITTLE"),
-      engineServer="host",
-      engineApiKey="apikey",
-      warnAboutDestructiveAction=False)
-
-    # Verify that deleteMetric was called only on the positives
-    expectedDeleteMetricCalls = [
-      mock.call(host="host", apiKey="apikey", metricName=metric)
-      for metric in positives
-    ]
-
-    self.maxDiff = None
-    self.assertItemsEqual(deleteMetricMock.call_args_list,
-                          expectedDeleteMetricCalls)
-
-    self.assertEqual(engineFactoryMock.return_value.begin.return_value
-                     .__enter__.return_value.execute.call_count, 2)
-
-
-
-  @patch("taurus.metric_collectors.metric_utils.collectorsdb.engineFactory",
-         autospec=True)
-  @patch("taurus.metric_collectors.metric_utils.CompanyDeleter"
-         "._flushTaurusEngineMetricDataPath",
-         spec_set=metric_utils.CompanyDeleter._flushTaurusEngineMetricDataPath)
-  @patch("taurus.metric_collectors.metric_utils.CompanyDeleter"
-         "._warnAboutDestructiveAction",
-         spec_set=metric_utils.CompanyDeleter._warnAboutDestructiveAction)
-  @patch("taurus.metric_collectors.metric_utils.getAllMetricSecurities",
-         autospec=True)
-  @patch("taurus.metric_collectors.metric_utils.getAllCustomMetrics",
-         autospec=True)
-  @patch("taurus.metric_collectors.metric_utils.deleteMetric",
-         autospec=True)
-  def testDeleteCompaniesWithWarningConfirmation(
-      self,
-      deleteMetricMock,
-      getAllCustomMetricsMock,
-      getAllMetricSecuritiesMock,
-      warnAboutDestructiveActionMock,
-      _flushTaurusEngineMetricDataPathMock,
-      engineFactoryMock):
-
-    negatives = set([
-      "TWITTER.TWEET.HANDLE.ZZZ.VOLUME",
-    ])
-
-    positives = set([
-      "XIGNITE.FOOBAR.CLOSINGPRICE",
-    ])
-
-    # Patch getAllCustomMetrics to return all negatives and positives
-    getAllCustomMetricsMock.return_value = [
-      {"name": metric} for metric in negatives.union(positives)
-    ]
-
-    engineFactoryMock.return_value.begin.return_value.__enter__.return_value \
-      .execute.return_value.rowcount = 1
-
-    getAllMetricSecuritiesMock.return_value = [
-      ("IBM", "exg"),
-      ("T", "exg")
-    ]
-
-    # Execute the function under test with default warnAboutDestructiveAction
-    metric_utils.CompanyDeleter.deleteCompanies(
-      tickerSymbols=("FOOBAR", "DOLITTLE"),
-      engineServer="host",
-      engineApiKey="apikey")
-
-    # Verify that _warnAboutDestructiveAction was called
-    self.assertEqual(warnAboutDestructiveActionMock.call_count, 1)
-
-    # Verify that deleteMetric was called only on the positives
-    expectedDeleteMetricCalls = [
-      mock.call(host="host", apiKey="apikey", metricName=metric)
-      for metric in positives
-    ]
-
-    self.maxDiff = None
-    self.assertItemsEqual(deleteMetricMock.call_args_list,
-                          expectedDeleteMetricCalls)
-
-    self.assertEqual(engineFactoryMock.return_value.begin.return_value
-                     .__enter__.return_value.execute.call_count, 1)
-
-
-  @patch("taurus.metric_collectors.metric_utils.CompanyDeleter"
-         "._warnAboutDestructiveAction",
-         spec_set=metric_utils.CompanyDeleter._warnAboutDestructiveAction)
-  def testDeleteCompaniesWithWarningRejection(self,
-                                              warnAboutDestructiveActionMock):
-
-    warnAboutDestructiveActionMock.side_effect = (
-      metric_utils.UserAbortedOperation)
-
-    # Execute the function under test with default warnAboutDestructiveAction
-    with self.assertRaises(metric_utils.UserAbortedOperation):
-      metric_utils.CompanyDeleter.deleteCompanies(
-        tickerSymbols=("FOOBAR", "DOLITTLE"),
-        engineServer="host",
-        engineApiKey="apikey")
-
-    # Verify that _warnAboutDestructiveAction was called
-    self.assertEqual(warnAboutDestructiveActionMock.call_count, 1)
-
-
-  @patch("taurus.metric_collectors.metric_utils.CompanyDeleter"
-         "._warnAboutDestructiveAction",
-         spec_set=metric_utils.CompanyDeleter._warnAboutDestructiveAction)
-  def testDeleteCompaniesWithWarningTimeout(self,
-                                            warnAboutDestructiveActionMock):
-
-    warnAboutDestructiveActionMock.side_effect = (
-      metric_utils.WarningPromptTimeout)
-
-    # Execute the function under test with default warnAboutDestructiveAction
-    with self.assertRaises(metric_utils.WarningPromptTimeout):
-      metric_utils.CompanyDeleter.deleteCompanies(
-        tickerSymbols=("FOOBAR", "DOLITTLE"),
-        engineServer="host",
-        engineApiKey="apikey")
-
-    # Verify that _warnAboutDestructiveAction was called
-    self.assertEqual(warnAboutDestructiveActionMock.call_count, 1)
-
-
-  @patch("taurus.metric_collectors.metric_utils.CompanyDeleter"
-         "._waitForFlusherAndGarbageCollect",
-         spec_set=metric_utils.CompanyDeleter._flushTaurusEngineMetricDataPath)
-  @patch("taurus.metric_collectors.metric_utils.metricDataBatchWrite",
-         autospec=True)
-  def testFlushTaurusEngineMetricDataPath(self,
-                                          metricDataBatchWriteMock,
-                                          waitForFlusherAndGarbageCollectMock):
-
-    metric_utils.CompanyDeleter._flushTaurusEngineMetricDataPath(
-      engineServer="host",
-      engineApiKey="apikey")
-
-    # Verity putSample called
-    self.assertEqual(metricDataBatchWriteMock.return_value.__enter__ \
-                     .return_value.call_count, 1)
-
-    # Verify _waitForFlusherAndGarbageCollect call
-    waitForFlusherAndGarbageCollectMock.assert_called_once_with(
-      engineServer="host",
-      engineApiKey="apikey",
-      flusherMetricName=ANY)
-
-
-  @patch("taurus.metric_collectors.metric_utils.time.sleep", autospec=True)
-  @patch("taurus.metric_collectors.metric_utils.getAllCustomMetrics",
-         autospec=True)
-  @patch("taurus.metric_collectors.metric_utils.deleteMetric",
-         autospec=True)
-  def testWaitForFlusherAndGarbageCollect(self,
-                                          deleteMetricMock,
-                                          getAllCustomMetricsMock,
-                                          _timeSleepMock):
-    oldFlusherMetricName = (
-      metric_utils.CompanyDeleter._DATA_PATH_FLUSHER_METRIC_PREFIX + "0")
-
-    flusherMetricName = (
-      metric_utils.CompanyDeleter._DATA_PATH_FLUSHER_METRIC_PREFIX + "1")
-
-    getAllCustomMetricsResultGroups = [
-      [
-        {"name": "XIGNITE.FOOBAR.CLOSINGPRICE"}
-      ],
-
-      [
-        {"name": "XIGNITE.FOOBAR.CLOSINGPRICE"},
-
-        {"name": oldFlusherMetricName}
-      ],
-
-      [
-        {"name": "XIGNITE.FOOBAR.CLOSINGPRICE"},
-
-        {"name": flusherMetricName}
-      ],
-    ]
-
-    # Patch getAllCustomMetrics to return all negatives and positives
-    getAllCustomMetricsMock.side_effect = iter(getAllCustomMetricsResultGroups)
-
-    # Execute
-    metric_utils.CompanyDeleter._waitForFlusherAndGarbageCollect(
-      engineServer="host",
-      engineApiKey="apikey",
-      flusherMetricName=flusherMetricName)
-
-    # Validate
-
-    self.assertEqual(getAllCustomMetricsMock.call_count,
-                     len(getAllCustomMetricsResultGroups))
-
-    expectedDeleteMetricCalls = [
-      mock.call(host="host", apiKey="apikey", metricName=oldFlusherMetricName),
-      mock.call(host="host", apiKey="apikey", metricName=flusherMetricName)
-    ]
-
-    self.assertSequenceEqual(deleteMetricMock.mock_calls,
-                             expectedDeleteMetricCalls)
-
-
-  @patch("time.sleep", autospec=True)
-  @patch("time.time", autospec=True)
-  @patch("taurus.metric_collectors.metric_utils.getAllCustomMetrics",
-         autospec=True)
-  def testWaitForFlusherAndGarbageCollectWithFlusherNotFound(
-      self,
-      getAllCustomMetricsMock,
-      timeMock,
-      sleepMock):
-
-    time_test_utils.configureTimeAndSleepMocks(timeMock, sleepMock)
-
-    flusherMetricName = (
-      metric_utils.CompanyDeleter._DATA_PATH_FLUSHER_METRIC_PREFIX + "1")
-
-
-    # Patch getAllCustomMetrics to return all negatives and positives
-    getAllCustomMetricsMock.return_value = [
-      {"name": "XIGNITE.FOOBAR.CLOSINGPRICE"}
-    ]
-
-    # Execute and validate
-    with self.assertRaises(metric_utils.FlusherMetricNotFound):
-      metric_utils.CompanyDeleter._waitForFlusherAndGarbageCollect(
-        engineServer="host",
-        engineApiKey="apikey",
-        flusherMetricName=flusherMetricName)
-
-
-  @patch("__builtin__.raw_input", autospec=True)
-  @patch("taurus.metric_collectors.metric_utils.random.randint", autospec=True)
-  def testWarnAboutDestructiveActionConfirmed(self, randintMock, rawInputMock):
-    randintMock.return_value = 1
-    rawInputMock.return_value = "Yes-1"
-
-    metric_utils.CompanyDeleter._warnAboutDestructiveAction(timeout=10,
-                                                            tickerSymbols="FOO",
-                                                            engineServer="host")
-
-    self.assertEqual(rawInputMock.call_count, 1)
-
-
-  @patch("__builtin__.raw_input", autospec=True)
-  def testWarnAboutDestructiveActionRejected(self, rawInputMock):
-    # NOTE: rejection can be anything other than the expected input of
-    # "Yes-<randint>"
-    rawInputMock.return_value = "No"
-
-    with self.assertRaises(metric_utils.UserAbortedOperation):
-      metric_utils.CompanyDeleter._warnAboutDestructiveAction(
-        timeout=10,
-        tickerSymbols="FOO",
-        engineServer="host")
-
-
-  @patch("__builtin__.raw_input", autospec=True)
-  def testWarnAboutDestructiveActionTimedOut(self, rawInputMock):
-
-    # NOTE: py.test by default captures console output and patches console input
-    # such # that raw_input fails. Although not ideal, we have to patch
-    # raw_input with something else that blocks and will be interrupted by
-    # SIGINT.
-    rawInputMock.side_effect = lambda *args: time.sleep(10)
-
-    with self.assertRaises(metric_utils.WarningPromptTimeout):
-      metric_utils.CompanyDeleter._warnAboutDestructiveAction(
-        timeout=0.001,
-        tickerSymbols="FOO",
-        engineServer="host")
